Revert implicit casting
Looks like netstandard2.0 wasn't meant to receive that new c#9 goodie

+++ PKHeX.Core/Editing/CommonEdits.cs
@@ -252,7 +252,7 @@ namespace PKHeX.Core
             if (pk is IGigantamax c)
                 c.CanGigantamax = Set.CanGigantamax;
             if (pk is IDynamaxLevel d)
-                d.DynamaxLevel = d.CanHaveDynamaxLevel(pk) ? 10 : 0;
+                d.DynamaxLevel = d.CanHaveDynamaxLevel(pk) ? (byte)10 : (byte)0;
 
             pk.ClearRecordFlags();
             pk.SetRecordFlags(Set.Moves);
