+++ PKHeX.Core/PKM/XK3.cs
@@ -69,7 +69,7 @@ namespace PKHeX.Core
         public override uint PID { get => BigEndian.ToUInt32(Data, 0x28); set => BigEndian.GetBytes(value).CopyTo(Data, 0x28); }
         // 0x2A-0x2B Unknown
         // 0x2C-0x2F Battle Related
-        public bool Obedient { get => Data[0x30] == 1; set => Data[0x30] = value ? 1 : 0; }
+        public bool Obedient { get => Data[0x30] == 1; set => Data[0x30] = value ? (byte)1 : (byte)0; }
         // 0x31-0x32 Unknown
         public int EncounterInfo { get => Data[0x33]; set => Data[0x33] = (byte)value; }
